Created an endpoint to https://www.frankfurter.app/docs/#latest which proves to be very, very convenient and helpful (Remember to thank person on github).
Created Objects to accompany the created object.

Separated Async Task "Get News" to a separate dedicated file, with an interface created to link the service and the async task together so that once the async task is completed, data is cached.

Speculations: Make the entire program run off one activity with multiple fragments for efficiency and as a good practice & create charts based on time series of the AUD exchange rate?